refactor: calendar logic divide to hook (#52)
