+++ src/constants/common.ts
@@ -1,0 +1,1 @@
+export const WEEK_DAY = ["일", "월", "화", "수", "목", "금", "토"];
